Fix model parsing and assign spawn points

+++ Kepler-Server/src/main/java/org/alexdev/kepler/dao/mysql/GameDao.java
@@ -86,7 +86,7 @@ public class GameDao {
 
             while (resultSet.next()) {
                 spawns.add(new GameSpawn(resultSet.getInt("team_id"),  resultSet.getInt("map_id"), resultSet.getString("type"),
-                        resultSet.getInt("x"), resultSet.getInt("y"), resultSet.getDouble("z")));
+                        resultSet.getInt("x"), resultSet.getInt("y"), resultSet.getInt("z")));
             }
 
         } catch (Exception e) {
